+++ docs/docs/about/release-notes.md
@@ -1,5 +1,35 @@
 # Release notes
 
+## [v5.4.0](https://github.com/pnp/cli-microsoft365/releases/tag/v5.4.0)
+
+### New commands
+
+**Planner:**
+
+- [planner task checklistitem add](../cmd/planner/task/task-checklistitem-add.md) - adds a new checklist item to a Planner task [#3280](https://github.com/pnp/cli-microsoft365/issues/3280)
+- [planner task remove](../cmd/planner/task/task-remove.md) - removes a Microsoft Planner task [#3227](https://github.com/pnp/cli-microsoft365/issues/3227)
+
+**SharePoint Online:**
+
+- [spo roledefinition remove](../cmd/spo/roledefinition/roledefinition-remove.md) - removes the role definition from the specified site [#3272](https://github.com/pnp/cli-microsoft365/issues/3272)
+
+### Changes
+
+- fixed 'planner task get' [#3324](https://github.com/pnp/cli-microsoft365/issues/3324)
+- extended spo list view add command with viewQuery option [#1862](https://github.com/pnp/cli-microsoft365/issues/1862)
+- updated planner task list options docs [#3347](https://github.com/pnp/cli-microsoft365/issues/3347)
+- updated PowerShell comparison docs [#3332](https://github.com/pnp/cli-microsoft365/issues/3332)
+- added short options to planner bucket commands [#3343](https://github.com/pnp/cli-microsoft365/issues/3343)
+- fixed spo listitem roleinheritance break docs option description formatting [#3358](https://github.com/pnp/cli-microsoft365/issues/3358)
+- added support for upgrading SPFx projects to v1.15.0-rc.0 [#3376](https://github.com/pnp/cli-microsoft365/issues/3376)
+- fixes listTitle option values with single quote [#3357](https://github.com/pnp/cli-microsoft365/issues/3357)
+- fixed planner documentation inconsistency [#3338](https://github.com/pnp/cli-microsoft365/issues/3338)
+- fixed invalid ISO date pattern with 7 milliseconds digits [#3372](https://github.com/pnp/cli-microsoft365/issues/3372)
+- fixed documentation for spo file checkout [#3388](https://github.com/pnp/cli-microsoft365/pull/3388)
+- fixed issue with mkdocs serve because of pygments version [#3389](https://github.com/pnp/cli-microsoft365/issues/3389)
+- fixed spo listitem list options [#3361](https://github.com/pnp/cli-microsoft365/issues/3361)
+- added aadGroup utils class [#3267](https://github.com/pnp/cli-microsoft365/issues/3267)
+
 ## [v5.3.0](https://github.com/pnp/cli-microsoft365/releases/tag/v5.3.0)
 
 ### New commands